% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Inputpara}
\alias{Inputpara}
\title{Inputpara}
\format{
An object of class \code{data.frame} with 1 rows and 31 columns.
}
\usage{
Inputpara
}
\description{
Just test dataframe(Description)
}
\examples{
head(Inputpara)
}
\keyword{datasets}
