% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onemonth_exam}
\alias{onemonth_exam}
\title{onemonth_exam}
\format{
An object of class \code{data.frame} with 1488 rows and 18 columns.
}
\usage{
onemonth_exam
}
\description{
Just test dataframe(Description)
}
\examples{
head(onemonth_exam)
}
\keyword{datasets}
