% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRIPLEX_CW_Flux.R
\name{TRIPLEX_CW_Flux}
\alias{TRIPLEX_CW_Flux}
\title{Carbon-water coupled model
a carbon-water coupled model integrated soil moisture and vapor pressure deficit into the stomatal conductance sub-model to estimate net ecosystem production and evapotranspiration in forest ecosystems}
\usage{
TRIPLEX_CW_Flux(Input_variable, Input_parameter)
}
\arguments{
\item{Input_variable}{Input variables of model}

\item{Input_parameter}{Input parameters of model}
}
\value{
the charts of simulated result for NEP and ET at 30 min scale, and monthly variation of the input environmental factors
}
\description{
Carbon-water coupled model
a carbon-water coupled model integrated soil moisture and vapor pressure deficit into the stomatal conductance sub-model to estimate net ecosystem production and evapotranspiration in forest ecosystems
}
\examples{
library(rTRIPLEXCWFlux)
TRIPLEX_CW_Flux (Input_variable=Inputvariable,Input_parameter=Inputpara)

}
